+++ Server/Lib/Net/NetRawUDP.cpp
@@ -203,21 +203,21 @@ namespace Net {
 			hrErr = NetSystem::RecvFrom(GetSocket(), pOver);
 			switch ((uint32_t)hrErr)
 			{
-			case ResultCode::SUCCESS_FALSE:
+			case (uint32_t)ResultCode::SUCCESS_FALSE:
 				hr = ResultCode::E_NET_TRY_AGAIN;
 				goto Proc_End;// success
 				break;
-			case ResultCode::SUCCESS:
-			case ResultCode::E_NET_IO_PENDING:
-			case ResultCode::E_NET_TRY_AGAIN:
-			case ResultCode::E_NET_WOULDBLOCK:
+			case (uint32_t)ResultCode::SUCCESS:
+			case (uint32_t)ResultCode::E_NET_IO_PENDING:
+			case (uint32_t)ResultCode::E_NET_TRY_AGAIN:
+			case (uint32_t)ResultCode::E_NET_WOULDBLOCK:
 				hr = hrErr;
 				goto Proc_End;// success
 				break;
-			case ResultCode::E_NET_NETUNREACH:
-			case ResultCode::E_NET_CONNABORTED:
-			case ResultCode::E_NET_CONNRESET:
-			case ResultCode::E_NET_NETRESET:
+			case (uint32_t)ResultCode::E_NET_NETUNREACH:
+			case (uint32_t)ResultCode::E_NET_CONNABORTED:
+			case (uint32_t)ResultCode::E_NET_CONNRESET:
+			case (uint32_t)ResultCode::E_NET_NETRESET:
 				// some remove has problem with continue connection
 				netTrace(TRC_NETCTRL, "UDP Remote has connection error err={0:X8}, {1}", hrErr, pOver->NetAddr.From);
 				//break;
@@ -242,18 +242,18 @@ namespace Net {
 		hrErr = NetSystem::SendTo(GetSocket(), pSendBuffer);
 		switch ((uint32_t)hrErr)
 		{
-		case ResultCode::E_NET_TRY_AGAIN:
+		case (uint32_t)ResultCode::E_NET_TRY_AGAIN:
 			break;
-		case ResultCode::SUCCESS:
-		case ResultCode::E_NET_IO_PENDING:
-		case ResultCode::E_NET_WOULDBLOCK:
+		case (uint32_t)ResultCode::SUCCESS:
+		case (uint32_t)ResultCode::E_NET_IO_PENDING:
+		case (uint32_t)ResultCode::E_NET_WOULDBLOCK:
 			break;
-		case ResultCode::E_NET_CONNABORTED:
-		case ResultCode::E_NET_CONNRESET:
-		case ResultCode::E_NET_NETRESET:
-		case ResultCode::E_NET_NOTCONN:
-		case ResultCode::E_NET_NOTSOCK:
-		case ResultCode::E_NET_SHUTDOWN:
+		case (uint32_t)ResultCode::E_NET_CONNABORTED:
+		case (uint32_t)ResultCode::E_NET_CONNRESET:
+		case (uint32_t)ResultCode::E_NET_NETRESET:
+		case (uint32_t)ResultCode::E_NET_NOTCONN:
+		case (uint32_t)ResultCode::E_NET_NOTSOCK:
+		case (uint32_t)ResultCode::E_NET_SHUTDOWN:
 			// Send fail by connection close
 			// Need to disconnect
 			hr = ResultCode::E_NET_CONNECTION_CLOSED;
@@ -404,20 +404,20 @@ namespace Net {
 		hrErr = NetSystem::RecvFrom(GetSocket(), pIOBuffer);
 		switch ((uint32_t)hrErr)
 		{
-		case ResultCode::SUCCESS_FALSE:
+		case (uint32_t)ResultCode::SUCCESS_FALSE:
 			hr = ResultCode::E_NET_TRY_AGAIN;
 			break;
-		case ResultCode::SUCCESS:
-		case ResultCode::E_NET_IO_PENDING:
-		case ResultCode::E_NET_TRY_AGAIN:
-		case ResultCode::E_NET_WOULDBLOCK:
+		case (uint32_t)ResultCode::SUCCESS:
+		case (uint32_t)ResultCode::E_NET_IO_PENDING:
+		case (uint32_t)ResultCode::E_NET_TRY_AGAIN:
+		case (uint32_t)ResultCode::E_NET_WOULDBLOCK:
 			hr = hrErr;
 			goto Proc_End;// success, but nothing to read
 			break;
-		case ResultCode::E_NET_NETUNREACH:
-		case ResultCode::E_NET_CONNABORTED:
-		case ResultCode::E_NET_CONNRESET:
-		case ResultCode::E_NET_NETRESET:
+		case (uint32_t)ResultCode::E_NET_NETUNREACH:
+		case (uint32_t)ResultCode::E_NET_CONNABORTED:
+		case (uint32_t)ResultCode::E_NET_CONNRESET:
+		case (uint32_t)ResultCode::E_NET_NETRESET:
 			// some remove has problem with connection
 			netTrace(TRC_NETCTRL, "UDP Remote has connection error err={0:X8}, {1}", hrErr, pIOBuffer->NetAddr.From);
 		default:
@@ -444,8 +444,8 @@ namespace Net {
 		{
 			switch ((uint32_t)hrRes)
 			{
-			case ResultCode::E_NET_CONNECTION_CLOSED:
-			case ResultCode::E_NET_IO_ABORTED:
+			case (uint32_t)ResultCode::E_NET_CONNECTION_CLOSED:
+			case (uint32_t)ResultCode::E_NET_IO_ABORTED:
 				hr = hrRes;
 				break;
 			default:
