+++ Server/Lib/Net/NetServerPeer.cpp
@@ -390,21 +390,21 @@ namespace Net {
 			hrErr = NetSystem::RecvFrom(GetSocket(), pOver);
 			switch ((uint32_t)hrErr)
 			{
-			case ResultCode::SUCCESS_FALSE:
+			case (uint32_t)ResultCode::SUCCESS_FALSE:
 				hr = ResultCode::E_NET_TRY_AGAIN;
 				goto Proc_End;// success
 				break;
-			case ResultCode::SUCCESS:
-			case ResultCode::E_NET_IO_PENDING:
-			case ResultCode::E_NET_TRY_AGAIN:
-			case ResultCode::E_NET_WOULDBLOCK:
+			case (uint32_t)ResultCode::SUCCESS:
+			case (uint32_t)ResultCode::E_NET_IO_PENDING:
+			case (uint32_t)ResultCode::E_NET_TRY_AGAIN:
+			case (uint32_t)ResultCode::E_NET_WOULDBLOCK:
 				hr = hrErr;
 				goto Proc_End;// success
 				break;
-			case ResultCode::E_NET_NETUNREACH:
-			case ResultCode::E_NET_CONNABORTED:
-			case ResultCode::E_NET_CONNRESET:
-			case ResultCode::E_NET_NETRESET:
+			case (uint32_t)ResultCode::E_NET_NETUNREACH:
+			case (uint32_t)ResultCode::E_NET_CONNABORTED:
+			case (uint32_t)ResultCode::E_NET_CONNRESET:
+			case (uint32_t)ResultCode::E_NET_NETRESET:
 				// some remove has problem with continue connection
 				netTrace(TRC_NETCTRL, "UDP Remote has connection error err={0:X8}, {1}", hrErr, pOver->NetAddr.From);
 				//break;
