+++ Server/Lib/Net/NetSystem_KQUEUE.cpp
@@ -149,13 +149,13 @@ namespace Net {
 				hr = hrErr;
 				switch ((uint32_t)hrErr)
 				{
-				case ResultCode::E_NET_TRY_AGAIN:
-				case ResultCode::E_NET_WOULDBLOCK:
-				case ResultCode::SUCCESS_FALSE:
+				case (uint32_t)ResultCode::E_NET_TRY_AGAIN:
+				case (uint32_t)ResultCode::E_NET_WOULDBLOCK:
+				case (uint32_t)ResultCode::SUCCESS_FALSE:
 					// These are expected return code
 					hr = ResultCode::SUCCESS;
 					break;
-				case ResultCode::E_NET_IO_PENDING:
+				case (uint32_t)ResultCode::E_NET_IO_PENDING:
 					Assert(false);
 					break;
 				default:
@@ -164,7 +164,7 @@ namespace Net {
 						netTrace(TRC_NETSYS, "ERROR KQUEUE Recv fail events:{0:X8} hr:{1:X8}", events, hrErr);
 					}
 					// fallthru
-				case ResultCode::SUCCESS:
+				case (uint32_t)ResultCode::SUCCESS:
 					// toss data to working thread
 					if (pReadBuffer != nullptr)
 					{
